fix: engine corruption and ply initialization

+++ src/utils/log.cpp
@@ -16,7 +16,7 @@ void log_uci(const std::string &msg, std::mutex &uci_mutex){
 
 }
 
-void log_uci(const int depth, const int bestScore, const int nodesVisited,  Move pv, std::chrono::time_point<std::chrono::steady_clock> startTime, std::mutex &uci_mutex) {
+void log_uci(const int depth, const int bestScore, const long long nodesVisited,  Move pv, std::chrono::time_point<std::chrono::steady_clock> startTime, std::mutex &uci_mutex) {
 
     auto now = std::chrono::steady_clock::now();
     auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(now - startTime).count();
